box2d: add version 2.4.2, remove version 2.3.1
