--- recipes/box2d/2.4.x/conandata.yml
@@ -1,10 +1,0 @@
-sources:
-  "2.4.0":
-    url: "https://github.com/erincatto/Box2D/archive/v2.4.0.zip"
-    sha256: "6aebbc54c93e367c97e382a57ba12546731dcde51526964c2ab97dec2050f8b9"
-  "2.4.1":
-    url: "https://github.com/erincatto/box2d/archive/refs/tags/v2.4.1.tar.gz"
-    sha256: "d6b4650ff897ee1ead27cf77a5933ea197cbeef6705638dd181adc2e816b23c2"
-patches:
-  "2.4.0":
-    - patch_file: "patches/0001-install-and-allow-shared.patch"
