--- recipes/box2d/2.4.x/test_v1_package/CMakeLists.txt
@@ -1,11 +1,0 @@
-cmake_minimum_required(VERSION 3.1)
-project(test_package)
-
-include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
-conan_basic_setup(TARGETS)
-
-find_package(box2d REQUIRED CONFIG)
-
-add_executable(${PROJECT_NAME} test_package.cpp)
-target_link_libraries(${PROJECT_NAME} box2d::box2d)
-set_property(TARGET ${PROJECT_NAME} PROPERTY CXX_STANDARD 11)
