--- recipes/box2d/2.4.x/test_v1_package/test_package.cpp
@@ -1,16 +1,0 @@
-#include <iostream>
-#include "box2d/box2d.h"
-
-int main(void) {
-    b2Vec2 gravity(0.0f, -10.0f);
-    b2World world(gravity);
-
-    b2BodyDef groundBodyDef;
-    groundBodyDef.position.Set(0.0f, -10.0f);
-
-    b2Body* groundBody = world.CreateBody(&groundBodyDef);
-    b2PolygonShape groundBox;
-    groundBox.SetAsBox(50.0f, 10.0f);
-
-    return 0;
-}
