+++ recipes/box2d/all/conandata.yml
@@ -1,4 +1,15 @@
 sources:
-  "2.3.1":
-    url: "https://github.com/erincatto/Box2D/archive/v2.3.1.zip"
-    sha256: "8ddd947400c59bcf09774bd75787e6253098a0bd337f5c61dae38b91aea678cf"
+  "2.4.2":
+    url: "https://github.com/erincatto/box2d/archive/refs/tags/v2.4.2.tar.gz"
+    sha256: "85b9b104d256c985e6e244b4227d447897fac429071cc114e5cc819dae848852"
+  "2.4.1":
+    url: "https://github.com/erincatto/box2d/archive/refs/tags/v2.4.1.tar.gz"
+    sha256: "d6b4650ff897ee1ead27cf77a5933ea197cbeef6705638dd181adc2e816b23c2"
+  "2.4.0":
+    url: "https://github.com/erincatto/Box2D/archive/v2.4.0.zip"
+    sha256: "6aebbc54c93e367c97e382a57ba12546731dcde51526964c2ab97dec2050f8b9"
+patches:
+  "2.4.0":
+    - patch_file: "patches/0001-install-and-allow-shared.patch"
+      patch_description: "add install, allow shared build"
+      patch_type: "conan"
