+++ recipes/box2d/all/conanfile.py
@@ -1,68 +1,71 @@
 import os
-from conan import ConanFile, tools
-from conan.tools.cmake import CMakeToolchain, CMake, cmake_layout
+from conan import ConanFile
+from conan.tools.files import apply_conandata_patches, export_conandata_patches, get, rm, rmdir, copy
+from conan.tools.cmake import CMake, CMakeToolchain, cmake_layout
+from conan.tools.scm import Version
 
-required_conan_version = ">=1.46.0"
+required_conan_version = ">=1.53.0"
 
 
 class Box2dConan(ConanFile):
     name = "box2d"
-    license = "Zlib"
     description = "Box2D is a 2D physics engine for games"
-    topics = ("physics", "engine", "game development")
-    homepage = "http://box2d.org/"
+    license = "Zlib"
     url = "https://github.com/conan-io/conan-center-index"
+    homepage = "http://box2d.org/"
+    topics = ("physics-engine", "graphic", "2d", "collision")
+    package_type = "library"
     settings = "os", "compiler", "build_type", "arch"
-    options = {"shared": [True, False],
-               "fPIC": [True, False]}
-    default_options = {"shared": False, "fPIC": True,}
+    options = {
+        "shared": [True, False],
+        "fPIC": [True, False]
+    }
+    default_options = {
+        "shared": False,
+        "fPIC": True
+    }
+
+    def export_sources(self):
+        export_conandata_patches(self)
 
     def config_options(self):
         if self.settings.os == "Windows":
-            try:
-                del self.options.fPIC
-            except Exception:
-                pass
+            self.options.rm_safe("fPIC")
 
     def configure(self):
         if self.options.shared:
-            try:
-                del self.options.fPIC
-            except Exception:
-                pass
+            self.options.rm_safe("fPIC")
 
     def layout(self):
-        cmake_layout(self, src_folder="Box2D")
+        cmake_layout(self, src_folder="src")
 
     def source(self):
-        tools.files.get(self,
-                        **self.conan_data["sources"][self.version],
-                        strip_root=True)
+        get(self, **self.conan_data["sources"][self.version], destination=self.source_folder, strip_root=True)
 
     def generate(self):
         tc = CMakeToolchain(self)
-        tc.variables["BOX2D_BUILD_SHARED"] = self.options.shared
-        tc.variables["BOX2D_BUILD_STATIC"] = not self.options.shared
-        if self.settings.os == "Windows" and self.options.shared:
-            tc.variables["CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS"] = True
-        tc.variables["BOX2D_BUILD_EXAMPLES"] = False
+        tc.cache_variables["CMAKE_POLICY_DEFAULT_CMP0077"] = "NEW"
+        tc.variables["BOX2D_BUILD_TESTBED"] = False
+        tc.variables["BOX2D_BUILD_UNIT_TESTS"] = False
         tc.generate()
 
     def build(self):
+        apply_conandata_patches(self)
         cmake = CMake(self)
-        cmake.configure(build_script_folder="Box2D")
+        cmake.configure()
         cmake.build()
 
     def package(self):
-        tools.files.copy(self, "License.txt", src=os.path.join(self.source_folder, "Box2D"), dst=os.path.join(self.package_folder,"licenses"))
-        tools.files.copy(self, os.path.join("Box2D", "*.h"), src=os.path.join(self.source_folder, "Box2D"), dst=os.path.join(self.package_folder, "include"))
-        tools.files.copy(self, "*.lib", src=self.build_folder, dst=os.path.join(self.package_folder, "lib"), keep_path=False)
-        tools.files.copy(self, "*.dll", src=self.build_folder, dst=os.path.join(self.package_folder, "bin"), keep_path=False)
-        tools.files.copy(self, "*.so*", src=self.build_folder, dst=os.path.join(self.package_folder, "lib"), keep_path=False)
-        tools.files.copy(self, "*.dylib", src=self.build_folder, dst=os.path.join(self.package_folder, "lib"), keep_path=False)
-        tools.files.copy(self, "*.a", src=self.build_folder, dst=os.path.join(self.package_folder, "lib"), keep_path=False)
+        copy(self, "LICENSE", dst=os.path.join(self.package_folder, "licenses"), src=self.source_folder)
+        cmake = CMake(self)
+        cmake.configure()
+        cmake.install()
+        rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
+        rm(self, "*.pdb", self.package_folder, recursive=True)
 
     def package_info(self):
-        self.cpp_info.libs = ["Box2D"]
-        if self.settings.os in ("FreeBSD", "Linux"):
-            self.cpp_info.system_libs = ["m"]
+        self.cpp_info.names["cmake_find_package"] = "box2d"
+        self.cpp_info.names["cmake_find_package_multi"] = "box2d"
+        self.cpp_info.libs = ["box2d"]
+        if Version(self.version) >= "2.4.1" and self.options.shared:
+            self.cpp_info.defines.append("B2_SHARED")
