+++ recipes/box2d/all/test_package/conanfile.py
@@ -1,11 +1,13 @@
 import os
-from conan import ConanFile, tools
+from conan import ConanFile
+from conan.tools.build import can_run
 from conan.tools.cmake import CMake, cmake_layout
 
 
 class Box2DTestConan(ConanFile):
     settings = "os", "compiler", "build_type", "arch"
     generators = "CMakeDeps", "CMakeToolchain", "VirtualRunEnv"
+    test_type = "explicit"
 
     def layout(self):
         cmake_layout(self)
@@ -19,6 +21,6 @@ class Box2DTestConan(ConanFile):
         cmake.build()
 
     def test(self):
-        if not tools.build.cross_building(self):
+        if can_run(self):
             bin_path = os.path.join(self.cpp.build.bindirs[0], "test_package")
             self.run(bin_path, env="conanrun")
