--- recipes/box2d/all/test_v1_package/CMakeLists.txt
@@ -1,8 +1,0 @@
-cmake_minimum_required(VERSION 3.4)
-project(test_package)
-
-include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
-conan_basic_setup()
-
-add_executable(test_package ../test_package/test_package.cpp)
-target_link_libraries(test_package ${CONAN_LIBS})
