+++ recipes/box2d/config.yml
@@ -1,7 +1,7 @@
 versions:
-  "2.3.1":
+  "2.4.2":
+    folder: "all"
+  "2.4.1":
     folder: "all"
   "2.4.0":
-    folder: "2.4.x"
-  "2.4.1":
-    folder: "2.4.x"
+    folder: "all"
